// src/models/Death.ts
export class Death{
    id: number;
    firstname: string;
    lastname: string;
    dateOfBirth: string;
    placeOfBirth: String;
    dateOfDeath: string;
    placeOfDeath: String;
    burialDate: string;
    recordDate: string;
    sectionNumber: Number;
    serialNumber: Number;
    graveNumber: Number;

  }